Completed all the junit testcases.
