+++ src/test/java/com/readcsvfile/devapp/ReadcsvfileApplicationTests.java
@@ -1,17 +1,17 @@
 package com.readcsvfile.devapp;
 
-import org.junit.jupiter.api.Test;
+
+import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-class ReadcsvfileApplicationTests {
+public class ReadcsvfileApplicationTests {
 
     @Test
-    void contextLoads() {
+   public void contextLoads() {
     }
 
 }
